+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/fetch-strategy/index.js
