+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/with-legacy-get-props/index.test.js
@@ -4,25 +4,27 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {withReactQuery} from './withReactQuery'
+import {withLegacyGetProps} from './index'
 import {shallow} from 'enzyme'
 import React from 'react'
 
-describe('withReactQuery', function() {
+describe('withLegacyGetProps', function() {
     test('Renders correctly', () => {
         const Wrapped = () => <p>Hello world</p>
-        const Component = withReactQuery(Wrapped)
+        const Component = withLegacyGetProps(Wrapped)
         const wrapper = shallow(<Component locals={{}} />)
         expect(wrapper.html()).toContain('Hello world')
     })
 
     test(`Has working getInitializers method`, () => {
-        expect(withReactQuery({}).getInitializers().length).toBe(1)
-        expect(withReactQuery({getInitializers: () => ['xyz']}).getInitializers().length).toBe(2)
+        expect(withLegacyGetProps({}).getInitializers().length).toBe(1)
+        expect(withLegacyGetProps({getInitializers: () => ['xyz']}).getInitializers().length).toBe(
+            2
+        )
     })
 
     test(`Has working getHOCsInUse method`, () => {
-        expect(withReactQuery({}).getHOCsInUse().length).toBe(1)
-        expect(withReactQuery({getHOCsInUse: () => ['xyz']}).getHOCsInUse().length).toBe(2)
+        expect(withLegacyGetProps({}).getHOCsInUse().length).toBe(1)
+        expect(withLegacyGetProps({getHOCsInUse: () => ['xyz']}).getHOCsInUse().length).toBe(2)
     })
 })
