+++ packages/template-typescript-minimal/app/components/_app-config/index.js
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {withLegacyGetProps} from 'pwa-kit-react-sdk/ssr/universal/withLegacyGetProps'
-import {withReactQuery} from 'pwa-kit-react-sdk/ssr/universal/withReactQuery'
+import {withLegacyGetProps} from 'pwa-kit-react-sdk/ssr/universal/components/with-legacy-get-props'
+import {withReactQuery} from 'pwa-kit-react-sdk/ssr/universal/components/with-react-query'
 import AppConfig from 'pwa-kit-react-sdk/ssr/universal/components/_app-config'
 
 export default withReactQuery(withLegacyGetProps(AppConfig))
